Close????

+++ src/main/java/com/igor_sikorsky_ist_hub/tunik_oleksandr/cw1_information_system_of_work_search/dao/EmployeeDao.java
@@ -14,9 +14,9 @@ import java.util.Collection;
 public interface EmployeeDao {
     //CRUD
     void create(Employee employee);
-    Employee findById(Integer id);
+    Employee findById(Long id);
     Employee findByLogin(String email);
     Collection<Employee> findAll();
     void update(Employee employee);
-    void deleteById(Integer id);
+    void deleteById(Long id);
 }
