+++ src/main/java/com/igor_sikorsky_ist_hub/tunik_oleksandr/cw1_information_system_of_work_search/dao/VacancyDao.java
@@ -14,10 +14,10 @@ import java.util.Collection;
 public interface VacancyDao {
     //CRUD
     void create(Vacancy vacancy);
-    Vacancy findById(Integer id);
+    Vacancy findById(Long id);
     Collection<Vacancy> findAll();
     void update(Vacancy vacancy);
-    void deleteById(Integer id);
+    void deleteById(Long id);
 }
